Make VersionFeature into an interface
